requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python
numpy
tensorflow
Pillow
python-jose[cryptography]
